Package structure refactor
