+++ airo-tulip/airo_tulip/hardware/controllers/controller.py
@@ -1,7 +1,7 @@
 from typing import List
 
 import numpy as np
-from airo_tulip.structs import WheelParamVelocity, WheelConfig
+from airo_tulip.hardware.structs import WheelParamVelocity, WheelConfig
 
 
 class Controller:
