+++ airo-tulip/airo_tulip/hardware/controllers/velocity_platform_controller.py
@@ -4,9 +4,9 @@ from typing import List, Tuple
 
 import numpy as np
 import rerun as rr
-from airo_tulip.controllers.controller import Controller
-from airo_tulip.structs import Attitude2DType, PlatformLimits, WheelConfig, WheelParamVelocity
-from airo_tulip.util import clip, clip_angle, get_shortest_angle
+from airo_tulip.hardware.controllers.controller import Controller
+from airo_tulip.hardware.structs import Attitude2DType, PlatformLimits, WheelConfig, WheelParamVelocity
+from airo_tulip.hardware.util import clip, clip_angle, get_shortest_angle
 from airo_typing import Vector2DType
