+++ airo-tulip/airo_tulip/hardware/robile_platform.py
@@ -1,12 +1,12 @@
 from typing import List
 
 import pysoem
-from airo_tulip.ethercat import EC_STATE_OPERATIONAL, EC_STATE_SAFE_OP
-from airo_tulip.logging.monitor_rerun import RerunMonitorLogger
-from airo_tulip.peripheral_client import PeripheralClient
-from airo_tulip.platform_driver import PlatformDriver, PlatformDriverType
-from airo_tulip.platform_monitor import PlatformMonitor
-from airo_tulip.structs import WheelConfig
+from airo_tulip.hardware.ethercat import EC_STATE_OPERATIONAL, EC_STATE_SAFE_OP
+from airo_tulip.hardware.logging.monitor_rerun import RerunMonitorLogger
+from airo_tulip.hardware.peripheral_client import PeripheralClient
+from airo_tulip.hardware.platform_driver import PlatformDriver, PlatformDriverType
+from airo_tulip.hardware.platform_monitor import PlatformMonitor
+from airo_tulip.hardware.structs import WheelConfig
 from loguru import logger
